backend/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.3.0
python-dotenv>=1.0.0
sqlmodel>=0.0.20
aiosqlite>=0.19.0
email-validator>=2.0.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
celery>=5.3.0
redis>=5.0.0
psutil>=5.9.0
sentry-sdk[fastapi]>=1.40.0

# JWT & Security
pyjwt>=2.8.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Database drivers
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# ML/AI
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

# LLM Providers
openai>=1.50.0
anthropic>=0.39.0
tiktoken>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
faker>=19.0.0

# Database migrations
alembic>=1.12.0

# OAuth2 & SSO
authlib>=1.2.0
python-multipart>=0.0.6

# MFA/2FA
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0

# Monitoring & Observability
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-prometheus>=0.41b0

# Message Queue
kombu>=5.3.0
pika>=1.3.0

# File Storage
boto3>=1.28.0
python-magic>=0.4.27

# PDF Generation
reportlab>=4.0.0
weasyprint>=59.0

# Websockets
websockets>=11.0.0

# GraphQL Support
strawberry-graphql[fastapi]>=0.216.0

# Machine Learning & Analytics
scipy>=1.11.0
pandas>=2.0.0

# Advanced AI Frameworks
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
pydantic-ai>=0.0.14
pydantic-ai-slim>=0.0.14
mem0ai>=0.1.0
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
reflex>=0.6.0
reflex-chakra>=0.6.0

